Technology
Created the technology mapping. This will be used for storing all
technologies. In the ITApplication you'll be able to link multiple
technologies to the application.

- Added Technology
- Added TechnologyController
- Added TechnologyDAL
- Added TechnologyDto
- Added TechnologyService
- Added TechnologyServiceImpl

- Changed small irregularities in ITApplication classes

+++ backend/src/main/java/com/bavostepbros/leap/domain/service/itapplicationservice/ITApplicationService.java
@@ -14,21 +14,21 @@ public interface ITApplicationService {
 
     ITApplication save(ITApplication itApplication);
 
-    ITApplication get(long itApplicationID);
+    ITApplication get(Integer itApplicationID);
 
     List<ITApplication> get(String name);
 
     List<ITApplication> getAll();
 
-    ITApplication update(long id, Integer statusID, String name, String technology, String version, LocalDate purchaseDate,
+    ITApplication update(Integer id, Integer statusID, String name, String technology, String version, LocalDate purchaseDate,
                          LocalDate endOfLife, Integer currentScalability, Integer expectedScalability,
                          Integer currentPerformance, Integer expectedPerformance, Integer currentSecurityLevel,
                          Integer expectedSecurityLevel, Integer currentStability, Integer expectedStability,
                          String costCurrency, String currentValue, Double currentYearlyCost, LocalDate timeValue);
 
-    void delete(long itApplicationID);
+    void delete(Integer itApplicationID);
 
-    boolean existsById(long itApplicationID);
+    boolean existsById(Integer itApplicationID);
 
     boolean existsByName(String name);
 }
